\name{check_data}
\alias{check_data}
\title{check genotype and phenotype data}
\usage{
Used to assess the availability of genotype and phenotype data.
}
\description{
Used to assess the availability of genotype and phenotype data, specifically including the number of rows and columns, the number of missing values, and the degree of match between genotypes and phenotypes.
}
\examples{
check_data(geno,pheno)
}
